pyproject.toml
[project]
name = "llm-twin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.40.54",
    "bs4>=0.0.2",
    "chromedriver-autoinstaller>=0.6.4",
    "datasets>=4.1.1",
    "google-generativeai>=0.8.5",
    "html2text>=2025.4.15",
    "langchain-community>=0.3.29",
    "langchain-google-genai>=2.0.10",
    "loguru>=0.7.3",
    "numpy>=2.3.2",
    "openai>=1.109.1",
    "opik>=1.8.77",
    "poethepoet>=0.37.0",
    "pymongo>=4.14.1",
    "qdrant-client>=1.15.1",
    "selenium>=4.32.0",
    "sentence-transformers>=5.1.0",
    "tqdm>=4.67.1",
    "transformers>=4.56.2",
    "typing-extensions>=4.15.0",
    "unsloth>=2025.7.2",
    "uvicorn>=0.35.0",
    "zenml[local,server]==0.90.0",
]

[tool.poe.tasks]
# Data pipelines
run-digital-data-etl-kd = "python -m tools.run --run-etl --no-cache --etl-config-filename digital_data_etl_kd.yaml"
run-digital-data-etl-maxime = "python -m tools.run --run-etl --no-cache --etl-config-filename digital_data_etl_maxime_labonne.yaml"
run-digital-data-etl-paul = "python -m tools.run --run-etl --no-cache --etl-config-filename digital_data_etl_paul_iusztin.yaml"
run-digital-data-etl = [
    "run-digital-data-etl-maxime",
    "run-digital-data-etl-paul",
]

run-feature-engineering-pipeline = "python -m tools.run --no-cache --run-feature-engineering"

run-end-to-end-data-pipeline = "python -m tools.run --no-cache --run-end-to-end-data"

